refactor: Rename init to bootstrap
