+++ docs/modules/ROOT/pages/index.adoc
@@ -35,26 +35,30 @@ To get started immediately, see xref:quickstart.adoc[Quickstart].
 
 The project follows a standard Rust project layout:
 
-
-* `src/`: Source code
-** `api/`: Route and controllers logic
-** `config/`: Configuration logic
-** `constants/`: Constant values used in the system
-** `domain/`: Domain logic
-** `init/`: Service initialization logic
-** `jobs/`: Asynchronous processing logic (queueing)
-** `logging/`: Logs and file rotation logic
-** `metrics/`: Metrics logic
-** `models/`: Data structures and types
-** `repositories/`: Configuration storage
-** `services/`: Services logic
-** `utils/`: Helper functions
-* `config/`: Configuration files
-* `tests/`: Integration tests
-* `docs/`: Documentation
-* `scripts/`: Utility scripts
-* `examples/`: Configuration examples
-* `helpers/`: Rust helper scripts
+```
+openzeppelin-relayer/
+├── src/
+│   ├── api/              # Route and controllers logic
+│   ├── bootstrap/        # Service initialization logic
+│   ├── config/           # Configuration logic
+│   ├── constants/        # Constant values used in the system
+│   ├── domain/           # Domain logic
+│   ├── jobs/             # Asynchronous processing logic (queueing)
+│   ├── logging/          # Logs File rotation logic
+│   ├── metrics/          # Metrics logic
+│   ├── models/           # Data structures and types
+│   ├── repositories/     # Configuration storage
+│   ├── services/         # Services logic
+│   └── utils/            # Helper functions
+│
+├── config/               # Configuration files
+├── tests/                # Integration tests
+├── docs/                 # Documentation
+├── scripts/              # Utility scripts
+├── examples/             # Configuration examples
+├── helpers/              # Rust helper scripts
+└── ... other root files (Cargo.toml, README.md, etc.)
+```
 
 
 For detailed information about each directory and its contents, see xref:structure.adoc[Project Structure Details].
