+++ src/bootstrap/initialize_app_state.rs
